requirements.txt
playwright>=1.40.0
ddddocr==1.0.6
requests
pillow>=10.0.0